Rename matplotlib_future_mode.py → matplotlib_compat.py
update all import sites accordingly
