+++ deeplabcut/core/visualization.py
@@ -17,7 +17,7 @@ from pathlib import Path
 import matplotlib.pyplot as plt
 import numpy as np
 
-from deeplabcut.utils.matplotlib_future_mode import get_colormap
+from deeplabcut.utils.matplotlib_compat import get_colormap
 
 
 def form_figure(nx, ny) -> tuple[plt.Figure, plt.Axes]:
