+++ deeplabcut/utils/visualization.py
@@ -33,7 +33,7 @@ from skimage import color, io
 from tqdm import trange
 
 from deeplabcut.utils import auxfun_videos, auxiliaryfunctions
-from deeplabcut.utils.matplotlib_future_mode import get_colormap
+from deeplabcut.utils.matplotlib_compat import get_colormap
 
 
 def get_cmap(n: int, name: str = "hsv") -> Colormap:
